NavBadges dissolves into user-menu chrome; admin dashboard aggregator moves to Users (nobodies-collective/Humans#1091)
- New ChromeSlots.UserMenu; _LoginPartial renders the slot where the hardcoded
  Issues link + NavBadges invocation was
- Humans.Issues contributes IssuesUserMenu (link + badge from its own service);
  IIssuesServiceRead deleted from the contracts leaf — its only external caller
  was NavBadges, so the count method is internal-only now
- IAdminDashboardService (minus GetPendingReviewCountAsync) + records move to
  Humans.Base.Interfaces.Admin; implementation to Humans.Users, registered in
  its Section
- NavBadgesViewComponent + view + gallery card deleted
