+++ src/Humans.Base/Resources/SharedResource.ca.resx
@@ -124,9 +124,6 @@
   <data name="Nav_OnboardingReview" xml:space="preserve">
     <value>Cua de revisió</value>
   </data>
-  <data name="Nav_Issues" xml:space="preserve">
-    <value>Incidències</value>
-  </data>
   <data name="CookieBanner_Title" xml:space="preserve">
     <value>Avís de galetes</value>
   </data>
