+++ src/Humans.Base/Resources/SharedResource.de.resx
@@ -124,9 +124,6 @@
   <data name="Nav_OnboardingReview" xml:space="preserve">
     <value>Prüfwarteschlange</value>
   </data>
-  <data name="Nav_Issues" xml:space="preserve">
-    <value>Vorgänge</value>
-  </data>
   <data name="CookieBanner_Title" xml:space="preserve">
     <value>Cookie-Hinweis</value>
   </data>
