+++ src/Humans.Base/Resources/SharedResource.es.resx
@@ -124,9 +124,6 @@
   <data name="Nav_OnboardingReview" xml:space="preserve">
     <value>Cola de revisión</value>
   </data>
-  <data name="Nav_Issues" xml:space="preserve">
-    <value>Incidencias</value>
-  </data>
   <data name="CookieBanner_Title" xml:space="preserve">
     <value>Aviso de cookies</value>
   </data>
