--- src/Humans.Web/Services/Dashboard/AdminDashboardData.cs
@@ -1,20 +1,0 @@
-namespace Humans.Web.Services.Dashboard;
-
-public record AdminDashboardData(
-    int TotalMembers,
-    int IncompleteSignup,
-    int PendingApproval,
-    int ActiveMembers,
-    int MissingConsents,
-    int Suspended,
-    int PendingDeletion,
-    int PendingApplications,
-    int TotalApplications,
-    int ApprovedApplications,
-    int RejectedApplications,
-    int ColaboradorApplied,
-    int AsociadoApplied,
-    IReadOnlyList<LanguageCount> LanguageDistribution,
-    UserSetMembership SetMembership);
-
-public record LanguageCount(string Language, int Count);
