--- src/Humans.Web/Views/Shared/Components/NavBadges/Default.cshtml
@@ -1,5 +1,0 @@
-@model int
-@if (Model > 0)
-{
-    <span class="badge rounded-pill nav-badge">@Model</span>
-}
