+++ src/Sections/Humans.Issues.Contracts/Humans.Issues.Contracts.csproj
@@ -14,10 +14,13 @@
     Humans.Issues/Contracts/ next to the IIssuesRetention it drives. IIssuesRetention has
     no consumer outside the section today and could follow it in; that is a later pass.
 
-    It holds exactly what lives outside the section: the nav-badge count Shell's
-    NavBadgesViewComponent renders, the retention call the job drives, and IssueCategory.
-    Everything else Issues does — the triage surface, the thread, the list projection, the
-    mutation results — has no consumer outside the section and stays internal.
+    It holds exactly what lives outside the section: the retention call the job drives, and
+    IssueCategory. The nav-badge count (IIssuesServiceRead.GetActionableCountForViewerAsync)
+    left this leaf at nobodies-collective/Humans#1091 — its sole external caller, Shell's
+    NavBadgesViewComponent, dissolved into this section's own user-menu chrome contribution,
+    so the method moved onto the internal-only IIssuesService. Everything else Issues does —
+    the triage surface, the thread, the list projection, the mutation results — has no
+    consumer outside the section and stays internal.
 
     References Humans.Interfaces only — the bottom of the graph.
   -->
