+++ src/Sections/Humans.Issues/Docs/Issues.md
@@ -167,7 +167,7 @@ Two controllers serve this section:
 - **Cross-domain navs removed, not just `[Obsolete]`-marked:** `Issue.Reporter`, `.Assignee`, `.ResolvedByUser`, `IssueComment.SenderUser` no longer exist as nav properties (nobodies-collective/Humans#1188). The repository does not `.Include()` them (there's nothing to include); `IssuesService` resolves display data in memory via `IUserServiceRead.GetUserInfosAsync` (design-rules §6b). EF configures the cross-section FK/cascade behavior with `HasOne<User>()` and no nav reference — there are no `#pragma warning disable CS0618` blocks left in the EF configurations.
 - **Cross-section calls** — the public interfaces this section consumes: `IUserServiceRead`, `IUserEmailService`, `IRoleAssignmentService`, `IEmailService`, `INotificationService`, `IAuditLogService`, `INavBadgeCacheInvalidator`, `IIssuesBadgeCacheInvalidator`.
 - **Nav-badge cache invalidation** uses two invalidators: `INavBadgeCacheInvalidator` (global nav count) and `IIssuesBadgeCacheInvalidator` (per-viewer actionable count). `IssuesService` holds both; the architecture test pins `IIssuesBadgeCacheInvalidator` explicitly. `IssuesBadgeCacheInvalidator` (impl in `Humans.Base/Caching/MemoryCacheInvalidators.cs`) — `MemoryCacheExtensions.InvalidateNavBadgeCounts()` drops every badge count by the same `CacheKeys` constant — is registered Scoped from this section's own `Section.Register`.
-- **Cross-section surface** — `Humans.Issues.Contracts` carries `IIssuesServiceRead.GetActionableCountForViewerAsync` (Shell's `NavBadgesViewComponent`), `IIssuesRetention.PurgeExpiredAsync` (driven by `CleanupIssuesJob`, which moved from `Humans.Infrastructure/Jobs` into this section's `Jobs/` at G5 lane 5b-1, then out of `Contracts/` entirely in the HUM0034 Jobs carve-out) and `IssueCategory` (named by `Humans.Agent`'s `route_to_issue` proposal). Everything else is internal to the section.
+- **Cross-section surface** — `Humans.Issues.Contracts` carries `IIssuesRetention.PurgeExpiredAsync` (driven by `CleanupIssuesJob`, which moved from `Humans.Infrastructure/Jobs` into this section's `Jobs/` at G5 lane 5b-1, then out of `Contracts/` entirely in the HUM0034 Jobs carve-out) and `IssueCategory` (named by `Humans.Agent`'s `route_to_issue` proposal). `IIssuesServiceRead.GetActionableCountForViewerAsync` was removed (nobodies-collective/Humans#1091): its sole caller, Shell's `NavBadgesViewComponent`, dissolved into this section's own `IssuesUserMenuViewComponent` chrome contribution, so the method moved onto the internal-only `IIssuesService`. Everything else is internal to the section.
 - **Architecture test** — `tests/Humans.Issues.Tests/Architecture/IssuesArchitectureTests.cs` pins the shape (constructor takes no EF type, the section localizes only through `IssuesResource`, audit discriminators are literals).
 
 ### Touch-and-clean guidance
