+++ src/Sections/Humans.Issues/IssuesResource.es.resx
@@ -46,6 +46,9 @@
   <data name="Issue_PageTitle" xml:space="preserve">
     <value>Incidencias</value>
   </data>
+  <data name="Issue_NavLabel" xml:space="preserve">
+    <value>Incidencias</value>
+  </data>
   <data name="Issue_New" xml:space="preserve">
     <value>Nueva incidencia</value>
   </data>
