+++ src/Sections/Humans.Issues/IssuesResource.it.resx
@@ -46,6 +46,9 @@
   <data name="Issue_PageTitle" xml:space="preserve">
     <value>Segnalazioni</value>
   </data>
+  <data name="Issue_NavLabel" xml:space="preserve">
+    <value>Segnalazioni</value>
+  </data>
   <data name="Issue_New" xml:space="preserve">
     <value>Nuova segnalazione</value>
   </data>
