+++ src/Sections/Humans.Issues/IssuesResource.resx
@@ -37,6 +37,7 @@
   <!-- ======================== -->
   <data name="Issue_Unavailable" xml:space="preserve"><value>That issue isn't available — it may have been removed, or you don't have access to it.</value></data>
   <data name="Issue_PageTitle" xml:space="preserve"><value>Issues</value></data>
+  <data name="Issue_NavLabel" xml:space="preserve"><value>Issues</value></data>
   <data name="Issue_New" xml:space="preserve"><value>New issue</value></data>
   <data name="Issue_New_Title" xml:space="preserve"><value>File a new issue</value></data>
   <data name="Issue_BackToList" xml:space="preserve"><value>Back to list</value></data>
