+++ tests/Humans.Issues.Tests/Architecture/IssuesArchitectureTests.cs
@@ -26,7 +26,7 @@ public class IssuesArchitectureTests
         var paramTypes = ctor.GetParameters().Select(p => p.ParameterType).ToList();
 
         paramTypes.Should().Contain(typeof(IIssuesBadgeCacheInvalidator),
-            because: "IssuesService owns the per-user actionable-count cache surfaced by NavBadgesViewComponent and must explicitly evict each affected viewer's entry on every count-shifting mutation (memory/code/viewcomponent-no-cache.md + code-review-rules.md §Cache Invalidation)");
+            because: "IssuesService owns the per-user actionable-count cache surfaced by IssuesUserMenuViewComponent and must explicitly evict each affected viewer's entry on every count-shifting mutation (memory/code/viewcomponent-no-cache.md + code-review-rules.md §Cache Invalidation)");
     }
 
     [HumansFact]
